+++ src/main/decorator/index.ts
@@ -1,6 +1,6 @@
-export * from './Controller';
-export * from './Handler';
-export * from './Injector';
-export * from './Method';
-export * from './Parameter';
-export * from './Service';
+export * from './controller.bak';
+export * from './handler.bak';
+export * from './injector.bak';
+export * from './method.bak';
+export * from './parameter.bak';
+export * from './service.bak';
