+++ src/main/decorator/service.bak.ts
@@ -1,4 +1,4 @@
-import { Injectable } from './Injector';
+import { Injectable } from './injector.bak';
 
 /**
  * Class Decorator.
